+++ packages/ui/src/lib/definitions/enums.ts
@@ -1,12 +1,6 @@
-export { ContentTypeEnum } from '@herodotus-cloud/core';
-export { SocialSourceEnum, CaptchaCategoryEnum, NotificationCategoryEnum } from '@herodotus-cloud/apis';
-
-export enum OperationEnum {
-  CREATE = 'create',
-  EDIT = 'edit',
-  AUTHORIZE = 'authorize',
-  INFO = 'info',
-  ALLOCATABLE = 'allocatable',
-  SETUP = 'setup',
-  INVOKE = 'invoke',
-}
+export { ContentTypeEnum, OperationEnum } from '@herodotus-cloud/core';
+export {
+  SocialSourceEnum,
+  CaptchaCategoryEnum,
+  NotificationCategoryEnum,
+} from '@herodotus-cloud/apis';