+++ shared/form-designer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @herodotus-cloud/form-designer
 
+## 1.7.0
+
+### Minor Changes
+
+- 版本发布
+
+### Patch Changes
+
+- Updated dependencies
+  - @herodotus-cloud/form-engine@1.7.0
+  - @herodotus-cloud/components@1.7.0
+  - @herodotus-cloud/form-apis@1.7.0
+  - @herodotus-cloud/core@1.7.0
+
 ## 1.6.6
 
 ### Patch Changes